reorganize functions in physics system
